requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apache-beam==2.55.1
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
crcmod==1.7
dataclasses-json==0.6.4
datasets==2.19.0
debugpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_563_nwtkoc/croot/debugpy_1690905063850/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.1.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastavro==1.9.4
fasteners==0.19
filelock==3.13.4
Flask==3.0.2
Flask-Executor==1.0.0
flask-redis==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1709821103657/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1708996616394/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1709559745751/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
Js2Py==0.74
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_782yoyc_98/croot/jupyter_core_1698937318631/work
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-experimental==0.0.57
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.27
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
multidict==6.0.5
multiprocess==0.70.16
mwparserfromhell==0.6.6
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==1.26.4
objsize==0.7.0
openai==1.12.0
orjson==3.9.15
packaging==23.2
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pinecone-client==3.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
proto-plus==1.23.0
protobuf==4.25.3
psutil @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1310b568-21f4-4cb0-b0e3-2f3d31e39728k9coaga5/croots/recipe/psutil_1656431280844/work
psycopg2-binary==2.9.9
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==14.0.2
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyjsparser==2.7.1
pymongo==4.6.3
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_43pxpbos3z/croot/pyzmq_1705605108344/work
redis==5.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
simple-websocket==1.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tenacity==8.2.3
tiktoken==0.5.2
tornado @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3a5nrn2jeh/croot/tornado_1696936974091/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1710254411456/work
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
zstandard==0.22.0
